refactor: KeyEventDispatcher 분리 #62
