+++ src/main/java/org/teamseven/tetris/ui/game/IKeyInputFeedback.java
@@ -1,0 +1,5 @@
+package org.teamseven.tetris.ui.game;
+
+public interface IKeyInputFeedback {
+    void feedback();
+}
